separating markdown+fs related code (db) from the real API that use it, removing api code from endpoints.
