+++ src/routes/[slug].json.js
@@ -1,10 +1,4 @@
-import { getPosts } from '$lib/posts.js';
-
-const lookup = new Map();
-getPosts()
-  .forEach(post => {
-    lookup.set(post.slug, JSON.stringify(post));
-  });
+import { getPostForSlug } from '$lib/posts.js';
 
 /**
  * @type {import('@sveltejs/kit').RequestHandler}
@@ -13,13 +7,15 @@ export function get({ params }) {
   // the `slug` parameter is available because
   // this file is called [slug].json.js
   const { slug } = params;
-  if (lookup.has(slug)) {
+  const post = getPostForSlug(slug);
+
+  if (post) {
     return {
       status: 200,
       headers: {
         'Content-Type': 'application/json'
       },
-      body: lookup.get(slug),
+      body: JSON.stringify(post),
     };
   } else {
     return {
