+++ src/routes/index.json.js
@@ -1,9 +1,6 @@
-import { getPosts } from '$lib/posts.js';
+import { getPublishedPosts } from '$lib/posts.js';
 
-const contents = JSON.stringify(getPosts()
-  .filter(it => it.metadata.published == 'true')
-  .filter(p => p.slug.indexOf('future/') < 0 && p.slug.indexOf('alternate-reality/') < 0)
-  .map(post => {
+const contents = getPublishedPosts().map(post => {
   return {
     title: post.metadata.title,
     date: post.metadata.dateString,
@@ -11,7 +8,7 @@ const contents = JSON.stringify(getPosts()
     slug: post.slug,
     minutesToRead: post.minutesToRead,
   };
-}));
+});
 
 /**
  * @type {import('@sveltejs/kit').RequestHandler}
@@ -22,6 +19,6 @@ export function get({ params }) {
     headers: {
       'Content-Type': 'application/json'
     },
-    body: contents
+    body: JSON.stringify(contents)
   };
 }
